Fix a misunderstanding.
Images should not be in this package... unless they are TRUELY common/universal. (which none of the mobile apps are)

We get google fonts from Google. Delete the offline-copies.

Ensure the direct compile of "all.scss" produces an empty string. 
This package is intended to provide variables and mixins for common patterns.